systemctl: don't unlink non-existing temporary files

+++ src/systemctl/systemctl-edit.c
@@ -627,6 +627,7 @@ int verb_edit(int argc, char *argv[], void *userdata) {
                         goto end;
                 }
 
+                f->tmp = mfree(f->tmp);
                 log_info("Successfully installed edited file '%s'.", f->path);
         }
 
@@ -640,7 +641,9 @@ int verb_edit(int argc, char *argv[], void *userdata) {
 
 end:
         for (EditFile *f = edit_files; f && f->path; f++) {
-                (void) unlink(f->tmp);
+
+                if (f->tmp)
+                        (void) unlink(f->tmp);
 
                 /* Removing empty dropin dirs */
                 if (!arg_full) {
